@@ -1,48 +1,55 @@
-import React, { useState } from "react";
-import { Button } from '@mui/material';
+import React from "react";
 import { Slide } from 'react-slideshow-image';
-import BgPic from "../../assets/bg.png"
+import { VideoWrapper } from "../../components/Page/Page";
 
-import {
-    VideoWrapper
-} from "../../components/Page/Page";
+// Define types for the props
+interface DropdownSlideProps {
+    slides: string[]; // Array of image URLs (strings)
+    videoLink: string; // URL for the video link
+}
 
-const DropdownSlide = ({ slides, videoLink }) => {
-
-    return (<Slide
-        autoplay={false}
-        onChange={function noRefCheck() { }}
-        onStartChange={function noRefCheck() { }}
-        easing={'ease'}
-        transitionDuration={300}>
-        {slides.map((imageUrl, index) => (
-            <div className="each-slide-effect" key={index}>
-                <div style={{
-                    backgroundImage: `url(${imageUrl})`,
-                    height: '70vh', // Adjust the height as needed
-                    backgroundSize: 'cover',
-                    backgroundPosition: 'center'
-                }}>
+const DropdownSlide: React.FC<DropdownSlideProps> = ({ slides, videoLink }) => {
+    return (
+        <Slide
+            autoplay={false}
+            onChange={function noRefCheck() { }}
+            onStartChange={function noRefCheck() { }}
+            easing={'ease'}
+            transitionDuration={300}
+        >
+            {slides.map((imageUrl: string, index: number) => (
+                <div className="each-slide-effect" key={index}>
+                    <div
+                        style={{
+                            backgroundImage: `url(${imageUrl})`,
+                            height: '70vh', // Adjust the height as needed
+                            backgroundSize: 'cover',
+                            backgroundPosition: 'center'
+                        }}
+                    >
+                    </div>
+                </div>
+            ))}
+            <div className="each-slide-effect">
+                <div
+                    style={{
+                        height: '70vh', // Adjust the height as needed
+                        backgroundSize: 'cover',
+                        backgroundPosition: 'center'
+                    }}
+                >
+                    <VideoWrapper>
+                        <iframe
+                            src={videoLink}
+                            title="YouTube video"
+                            allow="accelerometer; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
+                            allowFullScreen
+                        ></iframe>
+                    </VideoWrapper>
                 </div>
             </div>
-        ))}
-        <div className="each-slide-effect">
-            <div style={{
-                height: '70vh', // Adjust the height as needed
-                backgroundSize: 'cover',
-                backgroundPosition: 'center'
-            }}>
-                <VideoWrapper>
-                    <iframe
-                        src={videoLink}
-                        title="YouTube video"
-                        allow="accelerometer; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
-                        allowFullScreen
-                    ></iframe>
-                </VideoWrapper>
-            </div>
-        </div>
-    </Slide>)
+        </Slide>
+    );
 }
 
 export default DropdownSlide;
